+++ lib/groups/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './permissions';
 export * from './schema';
 export * from './assignments';
+export * from './classifications';